refactor: [고침] 리사이클러뷰/뷰페이저 어댑터 -> 어댑터 폴더로 이동
