+++ app/src/main/java/com/footprint/footprint/ui/adapter/HomeViewpagerAdapter.kt
@@ -1,7 +1,9 @@
-package com.footprint.footprint.ui.main.home
+package com.footprint.footprint.ui.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.footprint.footprint.ui.main.home.HomeDayFragment
+import com.footprint.footprint.ui.main.home.HomeMonthFragment
 
 class HomeViewpagerAdapter(fragment : Fragment) : FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = 2
