+++ app/src/main/java/com/footprint/footprint/ui/adapter/OnboardingViewpagerAdapter.kt
@@ -1,7 +1,8 @@
-package com.footprint.footprint.ui.onboarding
+package com.footprint.footprint.ui.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.footprint.footprint.ui.onboarding.*
 
 class OnboardingViewpagerAdapter(fragment: OnBoardingActivity): FragmentStateAdapter(fragment) {
     override fun getItemCount(): Int = 4
