+++ app/src/main/java/com/footprint/footprint/ui/register/RegisterActivity.kt
@@ -7,10 +7,10 @@ import androidx.recyclerview.widget.RecyclerView
 import com.footprint.footprint.data.model.UserModel
 import com.footprint.footprint.databinding.ActivityRegisterBinding
 import com.footprint.footprint.ui.BaseActivity
+import com.footprint.footprint.ui.adapter.RegisterViewpagerAdapter
 import com.footprint.footprint.ui.register.goal.RegisterGoalFragment
 import com.footprint.footprint.ui.register.info.RegisterInfoFragment
 import com.footprint.footprint.utils.getJwt
-import com.footprint.footprint.utils.saveJwt
 import com.google.android.material.tabs.TabLayoutMediator
